+++ clientprovider/src/test/java/org/mqnaas/clientprovider/impl/client/ClientProviderFactoryTest.java
@@ -7,7 +7,7 @@ import org.mqnaas.clientprovider.exceptions.ProviderNotFoundException;
 import org.mqnaas.clientprovider.impl.AbstractProviderFactory;
 import org.mqnaas.core.api.ICoreModelCapability;
 import org.mqnaas.core.api.IRootResource;
-import org.mqnaas.test.helpers.ReflectionTestHelper;
+import org.mqnaas.general.test.helpers.reflection.ReflectionTestHelper;
 import org.mqnaas.test.helpers.capability.ArtificialBundleGuard;
 import org.mqnaas.test.helpers.capability.TestCapabilitiesFactory;
 import org.mqnaas.test.helpers.clientprovider.EmptyClient;
