+++ test-helpers/src/test/java/org/mqnaas/general/test/helpers/TestClass.java
@@ -1,4 +1,4 @@
-package org.mqnaas.test.helpers;
+package org.mqnaas.general.test.helpers;
 
 /**
  * Class for testing purposes
